refactor(modularization): Phase 4 Step 4-S2c — marketplace DB split (CollectionMarketplace + UserCollectionSubscription + CollectionMarketplaceStatusEnum → aperag/domains/marketplace/db/models.py) (#25)
Third sub-step of Phase 4 Step 4-S2.

**New module** ``aperag/domains/marketplace/db/models.py``:
- ``CollectionMarketplace`` + ``UserCollectionSubscription``
  class bodies moved verbatim.
- ``CollectionMarketplaceStatusEnum`` moved alongside.
- Local ``_random_id`` helper.

**aperag/db/models.py**:
- Class bodies replaced with breadcrumbs.
- Re-export extended with marketplace symbols at the bottom.

**Verification**:
- ``collection_marketplace`` + ``user_collection_subscription``
  tables still on ``Base.metadata``.
- Boundary + Phase 3 re-export audit gates 17/17 green.
- ``ruff check`` + ``ruff format --check`` clean.
